+++ src/components/LibraryPickerItem.jsx
@@ -1,0 +1,30 @@
+var React = require('react');
+var cx = React.addons.classSet;
+
+var ProjectActions = require('../actions/ProjectActions');
+var config = require('../config');
+
+var LibraryPickerItem = React.createClass({
+  render: function() {
+    var library = this._getLibrary();
+    return (
+      <li className={cx({
+        'toolbar-menu-item': true,
+        'toolbar-menu-item--active': this.props.enabled
+      })} onClick={this._onClicked}>
+
+        {library.name}
+      </li>
+    );
+  },
+
+  _getLibrary: function() {
+    return config.libraries[this.props.libraryKey];
+  },
+
+  _onClicked: function() {
+    ProjectActions.toggleLibrary(this.props.projectKey, this.props.libraryKey);
+  }
+});
+
+module.exports = LibraryPickerItem;
