+++ src/components/Toolbar.jsx
@@ -1,15 +1,11 @@
 "use strict";
 
 var React = require('react');
-var cx = React.addons.classSet;
 var i18n = require('i18next-client');
-var _ = require('lodash');
-var moment = require('moment');
 
-var config = require('../config');
-var ProjectStore = require('../stores/ProjectStore');
+var LibraryPicker = require('./LibraryPicker.jsx');
 var ProjectActions = require('../actions/ProjectActions');
-var CurrentProjectActions = require('../actions/CurrentProjectActions');
+var ProjectList = require('./ProjectList.jsx');
 
 var Toolbar = React.createClass({
   getInitialState: function() {
@@ -101,118 +97,6 @@ var Toolbar = React.createClass({
         return {open: true};
       }
     });
-  }
-});
-
-var LibraryPicker = React.createClass({
-  getInitialState: function() {
-    return this._calculateState();
-  },
-
-  componentDidMount: function() {
-    ProjectStore.addChangeListener(this._onChange);
-  },
-
-  componentWillUnmount: function() {
-    ProjectStore.removeChangeListener(this._onChange);
-  },
-
-  render: function() {
-    var libraries = _.map(config.libraries, function(library, key) {
-      return (
-        <LibraryPickerItem
-          projectKey={this.props.projectKey}
-          libraryKey={key}
-          enabled={this._isLibraryEnabled(key)} />
-      );
-    }.bind(this));
-
-    return <ul className="toolbar-menu">{libraries}</ul>;
-  },
-
-  _onChange: function() {
-    this.setState(this._calculateState());
-  },
-
-  _calculateState: function() {
-    var enabledLibraries = [];
-    if (this.props.projectKey) {
-      enabledLibraries =
-        ProjectStore.get(this.props.projectKey).enabledLibraries;
-    }
-    return {enabledLibraries: enabledLibraries};
-  },
-
-  _isLibraryEnabled: function(libraryKey) {
-    return this.state.enabledLibraries.indexOf(libraryKey) !== -1;
-  }
-});
-
-var LibraryPickerItem = React.createClass({
-  render: function() {
-    var library = this._getLibrary();
-    return (
-      <li className={cx({
-        'toolbar-menu-item': true,
-        'toolbar-menu-item--active': this.props.enabled
-      })} onClick={this._onClicked}>
-
-        {library.name}
-      </li>
-    );
-  },
-
-  _getLibrary: function() {
-    return config.libraries[this.props.libraryKey];
-  },
-
-  _onClicked: function() {
-    ProjectActions.toggleLibrary(this.props.projectKey, this.props.libraryKey);
-  }
-});
-
-var ProjectList = React.createClass({
-  getInitialState: function() {
-    return this._calculateState();
-  },
-
-  componentDidMount: function() {
-    ProjectStore.addChangeListener(this._onChange);
-  },
-
-  componentWillUnmount: function() {
-    ProjectStore.removeChangeListener(this._onChange);
-  },
-
-  render: function() {
-    var projects = this.state.projects.map(function(project) {
-      return (
-        <li className='toolbar-menu-item'
-          onClick={this._onProjectClicked.bind(this, project)}>
-          <div>{moment(project.updatedAt).fromNow()}</div>
-          <div><code>{project.sources.html.slice(0, 50)}</code></div>
-          <div><code>{project.sources.css.slice(0, 50)}</code></div>
-          <div>
-            <code>{project.sources.javascript.slice(0, 50)}</code>
-          </div>
-        </li>
-      );
-    }.bind(this));
-
-    return <ul className="toolbar-menu">{projects}</ul>;
-  },
-
-  _onChange: function() {
-    this.setState(this._calculateState());
-  },
-
-  _calculateState: function() {
-    return {projects: ProjectStore.all()};
-  },
-
-  _onProjectClicked: function(project) {
-    CurrentProjectActions.select(project.projectKey);
-    this.props.onProjectSelected();
   }
 });
